Split up code into smaller files
