+++ obi_one/utils/circuit_registration/__init__.py
@@ -1,0 +1,65 @@
+"""Circuit registration utilities for entitycore."""
+
+from obi_one.utils.circuit_registration.assets import (
+    CIRCUIT_ASSET_MAPPING,
+    register_asset,
+)
+from obi_one.utils.circuit_registration.generate import (
+    generate_additional_circuit_assets,
+    generate_compressed_circuit_asset,
+    generate_connectivity_matrix_asset,
+    generate_connectivity_plot_assets,
+    generate_overview_image_asset,
+    generate_sim_designer_image_asset,
+)
+from obi_one.utils.circuit_registration.links import (
+    register_contributions,
+    register_derivation,
+    register_publication_links,
+)
+from obi_one.utils.circuit_registration.register import (
+    register_circuit,
+    register_circuit_from_metadata,
+)
+from obi_one.utils.circuit_registration.resolve import (
+    check_if_circuit_exists,
+    find_agent,
+    find_role,
+    get_brain_region,
+    get_circuit,
+    get_contributions,
+    get_exp_date,
+    get_license,
+    get_parent_circuit,
+    get_publications,
+    get_root_circuit,
+    get_subject,
+)
+
+__all__ = [
+    "CIRCUIT_ASSET_MAPPING",
+    "check_if_circuit_exists",
+    "find_agent",
+    "find_role",
+    "generate_additional_circuit_assets",
+    "generate_compressed_circuit_asset",
+    "generate_connectivity_matrix_asset",
+    "generate_connectivity_plot_assets",
+    "generate_overview_image_asset",
+    "generate_sim_designer_image_asset",
+    "get_brain_region",
+    "get_circuit",
+    "get_contributions",
+    "get_exp_date",
+    "get_license",
+    "get_parent_circuit",
+    "get_publications",
+    "get_root_circuit",
+    "get_subject",
+    "register_asset",
+    "register_circuit",
+    "register_circuit_from_metadata",
+    "register_contributions",
+    "register_derivation",
+    "register_publication_links",
+]
